TechProject/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.0
Flask==2.2.5
fsspec==2023.1.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
loguru==0.7.0
MarkupSafe==2.1.2
nltk==3.8.1
numpy==1.21.6
opencv-python==4.7.0.72
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
psutil==5.9.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
simplejson==3.19.1
six==1.16.0
stanfordcorenlp==3.9.1.1
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tqdm==4.65.0
transformers==4.29.1
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.2.3
zipp==3.15.0
